requirements.txt
pyotp==2.3.0
Pillow==7.0.0
